feat(226): done with binary tree inversion
this was actually a very elegant solution and gentle introduction to
binary trees
i only remembered the three types of traversal, but i'm starting to
remember the whys of each

i forgot this, but neetcode mentions that this is called depth first
search. we are trying to swap the left and right pointers for each node,
but left and right children's nodes need to be swapped as well.
- this is very naturally a recursive solution
- there also exists an iterative approach with a stack

there also exists the breadth first search algorithm where we maintain a
queue structure
- this is an iterative approach that uses a `while` loop