feat(3): done with problem 3 without _much_ help
okay fine i looked at the hint that says to generate all the substrings
this solution is def not as efficient as the map solution

the idea for the map solution is that we're keeping trak of the current
"window" using a dictionary. inside the dictionary, we'll map the
characters to the index at which they were last seen

if we encounter a character that is already in the map, then we can
increment the left index to where that encountered character already was
already seen

better than mine because mine just naiively increments by 1 and keeps
checking. not horrible, but meh i'm just glad that i figured it out
after scratching my head for an hour